Found an independent way of executing JMX SSL tests.
JMX SSL tests no longer need to be executed in separate JVM.
