+++ jqm-all/jqm-engine/src/main/java/com/enioka/jqm/tools/JmxSslHandshakeListener.java
@@ -210,4 +210,12 @@ public class JmxSslHandshakeListener implements HandshakeCompletedListener
         }
     }
 
+    /**
+     * Reset all data of {@link #sslSuccessClientsUsernames};
+     */
+    public void reset()
+    {
+        sslSuccessClientsUsernames.clear();
+    }
+
 }
